+++ browser-search-regex/chrome/custom_lib/DOM_manipulator/components/BSRInput/BSRInput.ts
@@ -1,39 +1,4 @@
-interface highlightMeOptions {
-  color: string;
-  mods: string;
-  limit: number;
-}
-interface nextMatchOptions {
-  direction: number;
-  newStyles: any;
-  oldStyles: any;
-  scrollBehavior: 'smooth' | 'auto';
-  scrollable: boolean;
-}
 namespace Components {
-  export class BSRPopupCard extends HTMLElement {
-    constructor() {
-      super();
-      const shadowRoot = this.attachShadow({ mode: 'open' });
-
-      const bsrPopupCard = document.createElement('div');
-      bsrPopupCard.className = 'BSRPopupWrapper shadowWrapper';
-      bsrPopupCard.innerHTML = `
-        <div id="bsr-control-wrapper">
-          <span id="bsr-input-button" class="BSRButton BSRControlButton">NEW</span>
-          <span id="bsr-exit-button" class="BSRButton BSRControlButton">X</span>
-        </div>
-        <bsr-break></bsr-break>
-        <div id="bsr-form-wrapper"></div>
-      `;
-      Globals.formWrapper = bsrPopupCard.querySelector('#bsr-form-wrapper');
-
-      const style = document.createElement('style');
-      style.textContent = styles;
-      shadowRoot.appendChild(style);
-      shadowRoot.appendChild(bsrPopupCard);
-    }
-  }
   export class BSRInput extends HTMLElement {
     public preserveCase = 'i';
 
@@ -128,6 +93,7 @@ namespace Components {
     public inputWrapper: HTMLElement;
 
     public colorCopyButton: HTMLElement;
+
     public colorCopyTooltip: HTMLElement;
 
     /**
@@ -138,60 +104,6 @@ namespace Components {
       super();
       const shadowRoot = this.attachShadow({ mode: 'open' });
 
-      const bsrInput = document.createElement('div');
-      bsrInput.className = 'BSRInputWrapper shadowWrapper';
-      bsrInput.innerHTML = `
-        <div class="BSRInputTopHalf">
-          <input class="BSRMainInputField" type="text" placeholder="regular expression" name="some-key"/>
-          <span class="BSRModifierWrapper">
-            <div class="BSRButton BSRModifierCoverButton">V</div>
-            <div class="BSRModifierDropdown">
-              <div class="BSRScrollBar">
-                <div class="BSRButton BSRModifierButton">
-                  <input class="BSRModifierInput" id="bsr-exact-match" type="checkbox" checked />
-                  <label for="bsr-exact-match">Exact match</label>
-                </div>
-                <div class="BSRButton BSRModifierButton">
-                  <input class="BSRModifierInput" id="bsr-is-regex" type="checkbox" />
-                  <label for="bsr-is-regex">Regular expression</label>
-                </div>
-                <div class="BSRButton BSRModifierButton">
-                  <input class="BSRModifierInput" id="bsr-levenshtein" type="checkbox" />
-                  <label for="bsr-levenshtein">Loose search</label>
-                </div>
-                <div class="BSRButton BSRModifierButton">
-                  <input class="BSRModifierInput" id="bsr-case-sensitive" type="checkbox" />
-                  <label for="bsr-case-sensitive">Case sensitive</label>
-                </div>
-                <div class="BSRButton BSRModifierButton">
-                  <input class="BSRModifierInput" id="bsr-should-scroll" type="checkbox" />
-                  <label for="bsr-should-scroll">Stop auto scroll</label>
-                </div>
-                <div class="BSRButton BSRModifierButton BSRMaxMatchLimitWrapper">
-                  <input class="BSRModifierInput BSRMaxMatchLimit" id="bsr-max-matches" type="number" value="100"/>
-                  <div>Maximum matches</div>
-                </div>
-                <div class="BSRButton BSRModifierButton BSRColorPickerWrapper">
-                  <input class="BSRModifierInput BSRColorPicker" id="bsr-color-input" type="color" value="#FBFF00"/>
-                    <div>Selection color <span class="BSRButton BSRColorFacts"><span>#FBFF00</span> <span class="BSRCopyButton BSRColorCopyButton">⛶<div class="BSRToolTip BSRColorCopyToolTip">Copied</div></span></div>
-                </div>
-              </div>
-            </div>
-          </span>
-        </div>
-        <div class="BSRInputBottomHalf">
-          <span class="BSRButton BSRActionButton BSRPrevButton">⇐</span>
-          <span class="BSRButton BSRActionButton BSRNextButton">⇒</span>
-          <span class="BSRButton BSRActionButton BSRDeleteButton">-</span>
-          <span class="BSRButton BSRActionButton BSRCopyButton">⛶</span>
-          <span style="flex-grow:1;"></span>
-          <span class="BSRFoundMatches BSRActionButton">
-            <span class="BSRMatchNumerator">0</span>
-            /
-            <span class="BSRMatchDenominator">0</span>
-          </span>
-        </div>
-      `;
       const style = document.createElement('style');
       style.textContent = styles;
       shadowRoot.appendChild(style);
@@ -200,50 +112,67 @@ namespace Components {
 
       // create the key to identify this element
       this.key = `regex-key-${Math.random().toString(36).substring(2, 5)}`;
+
       this.inputWrapper = bsrInput;
+
       this.searchInput = bsrInput.querySelector(
         'input.BSRMainInputField'
       ) as HTMLInputElement;
+
       this.caseSensitive = bsrInput.querySelector(
         '#bsr-case-sensitive'
       ) as HTMLInputElement;
+
       this.exactMatch = bsrInput.querySelector(
         '#bsr-exact-match'
       ) as HTMLInputElement;
+
       this.isRegex = bsrInput.querySelector(
         '#bsr-is-regex'
       ) as HTMLInputElement;
+
       this.levenshtein = bsrInput.querySelector(
         '#bsr-levenshtein'
       ) as HTMLInputElement;
+
       this.shouldScroll = bsrInput.querySelector(
         '#bsr-should-scroll'
       ) as HTMLInputElement;
+
       this.maxMatchLimit = bsrInput.querySelector(
         '#bsr-max-matches'
       ) as HTMLInputElement;
+
       this.colorInput = bsrInput.querySelector(
         '#bsr-color-input'
       ) as HTMLInputElement;
+
       this.colorFacts = bsrInput.querySelector('.BSRColorFacts') as HTMLElement;
+
       this.next = bsrInput.querySelector('.BSRNextButton') as HTMLElement;
+
       this.prev = bsrInput.querySelector('.BSRPrevButton') as HTMLElement;
+
       this.countNum = bsrInput.querySelector(
         '.BSRMatchNumerator'
       ) as HTMLElement;
+
       this.countDen = bsrInput.querySelector(
         '.BSRMatchDenominator'
       ) as HTMLElement;
+
       this.minus = bsrInput.querySelector('.BSRDeleteButton') as HTMLElement;
+
       this.copy = bsrInput.querySelector('.BSRCopyButton') as HTMLElement;
+
       this.colorCopyButton = bsrInput.querySelector(
         '.BSRColorCopyButton'
       ) as HTMLElement;
+
       this.colorCopyTooltip = bsrInput.querySelector(
         '.BSRColorCopyToolTip'
       ) as HTMLElement;
 
-      // this.colorCopyButton.addEventListener('click', () => {});
       // auto focus the input
       setTimeout(() => this.searchInput.focus(), 1);
       // event listeners
@@ -278,6 +207,7 @@ namespace Components {
           hexCode.innerHTML = this.colorInput.value;
         }
       });
+
       this.colorFacts.addEventListener('click', () => {
         let hexCode = this.colorFacts.querySelector('span');
         if (hexCode != null) {
@@ -290,6 +220,7 @@ namespace Components {
           // this.colorCopyTooltip.style.display = 'none';
         }, 1000);
       });
+
       this.caseSensitive.addEventListener('change', (e) => {
         e.preventDefault();
         if (this.caseSensitive.checked) {
@@ -300,6 +231,23 @@ namespace Components {
         this.handleHighlighting();
         this.searchInput.focus();
       });
+
+      this.shouldScroll.addEventListener('change', (e) => {
+        e.preventDefault();
+        if (this.shouldScroll.checked) {
+          this.preserveScroll = false;
+        } else {
+          this.preserveScroll = true;
+        }
+        this.searchInput.focus();
+      });
+
+      this.maxMatchLimit.addEventListener('input', (e) => {
+        e.preventDefault();
+        this.maxLimit = Number(this.maxMatchLimit.value);
+        this.handleHighlighting();
+      });
+
       this.exactMatch.addEventListener('change', (e) => {
         e.preventDefault();
         if (this.exactMatch.checked) {
@@ -310,6 +258,7 @@ namespace Components {
         this.handleHighlighting();
         this.searchInput.focus();
       });
+
       this.isRegex.addEventListener('change', (e) => {
         e.preventDefault();
         if (this.isRegex.checked) {
@@ -320,6 +269,7 @@ namespace Components {
         this.handleHighlighting();
         this.searchInput.focus();
       });
+
       this.levenshtein.addEventListener('change', (e) => {
         e.preventDefault();
         if (this.levenshtein.checked) {
@@ -330,20 +280,7 @@ namespace Components {
         this.handleHighlighting();
         this.searchInput.focus();
       });
-      this.shouldScroll.addEventListener('change', (e) => {
-        e.preventDefault();
-        if (this.shouldScroll.checked) {
-          this.preserveScroll = false;
-        } else {
-          this.preserveScroll = true;
-        }
-        this.searchInput.focus();
-      });
-      this.maxMatchLimit.addEventListener('input', (e) => {
-        e.preventDefault();
-        this.maxLimit = Number(this.maxMatchLimit.value);
-        this.handleHighlighting();
-      });
+
       this.next.addEventListener('click', (e: any) => {
         e.preventDefault();
         this.nextOrPrev = this.next;
@@ -369,6 +306,7 @@ namespace Components {
         }
         this.searchInput.focus();
       });
+
       this.prev.addEventListener('click', (e: any) => {
         e.preventDefault();
         this.nextOrPrev = this.prev;
@@ -394,6 +332,7 @@ namespace Components {
         }
         this.searchInput.focus();
       });
+
       this.minus.addEventListener('click', (e) => {
         e.preventDefault();
         Highlighter.clearHighlight(this.key);
@@ -405,6 +344,7 @@ namespace Components {
         this.parentElement.removeChild(this);
         Globals.INPUT_AMT--;
       });
+
       this.copy.addEventListener('click', (e) => {
         e.preventDefault();
         let GI = Globals.getGI(this.key);
@@ -421,6 +361,11 @@ namespace Components {
         this.searchInput.focus();
       });
     }
+    // end of constructor
+
+    /**
+     * handles all highlighting
+     */
     handleHighlighting() {
       if (
         this.highlightMe(this.searchInput.value, this.searchType, {
@@ -437,6 +382,14 @@ namespace Components {
         this.countDen.innerHTML = '0';
       }
     }
+
+    /**
+     *
+     * @param searchTerm term to search for
+     * @param searchType the type of search
+     * @param options
+     * @returns boolean indicating if the highlight was carried out successfully
+     */
     highlightMe(
       searchTerm = '',
       searchType: 'exact' | 'regexp' | 'lev',
@@ -516,6 +469,14 @@ namespace Components {
       }
       return false;
     }
+
+    /**
+     * creates a tag to simulate the highlighting
+     * @param match the text to store inside the tag
+     * @param sameMatchID the id of the current node within the match
+     * @param color a color to highlight with
+     * @returns an element to insert into the dom
+     */
     createTag(match: string, sameMatchID: number, color: string) {
       var highlightMeElem = document.createElement('highlight-me');
       highlightMeElem.className = `chrome-bsr-highlight-me ${this.key}`;
@@ -530,6 +491,12 @@ namespace Components {
       highlightMeElem.textContent = match;
       return highlightMeElem;
     }
+
+    /**
+     * changes the color of every match given a key
+     * @param key the key of matches to change color
+     * @param color color to change to
+     */
     changeColor(key: string, color: string) {
       let matches = document.querySelectorAll(`highlight-me.${key}`);
       matches.forEach((elem: any) => {
@@ -538,6 +505,11 @@ namespace Components {
       });
     }
 
+    /**
+     * inverts a color to get a decently contrasting color
+     * @param hex a hex value
+     * @returns a hex value
+     */
     invertColor(hex: string) {
       if (hex.indexOf('#') === 0) {
         hex = hex.slice(1);
@@ -753,240 +725,4 @@ namespace Components {
       return null;
     }
   }
-  export const queryShadowSelector = (elem: HTMLElement, selector: string) => {
-    const shadow = elem.shadowRoot;
-    if (shadow != null) {
-      const childNodes = Array.from(shadow.childNodes);
-      let parent = childNodes.find((node) => {
-        if (
-          node instanceof HTMLElement &&
-          node.className.includes('shadowWrapper')
-        ) {
-          return node;
-        }
-      });
-      if (parent instanceof HTMLElement && parent != null) {
-        return parent.querySelector(selector);
-      }
-    }
-    return null;
-  };
 }
-const styles = `
-*{
-  font-family: 'Almarai';
-}
-/* no select */
-div {
-  user-select: none;
-}
-span {
-  user-select: none;
-}
-
-/* custom break */
-bsr-break {
-  display: block;
-  height: 0.5px;
-  margin: 0 5% 2% 5%;
-  background-color: #3c3f41;
-}
-
-/* scroll bars */
-.BSRScrollBar {
-  max-height: inherit;
-  overflow-y: scroll;
-}
-.BSRScrollBar::-webkit-scrollbar {
-  width: 4px;
-}
-.BSRScrollBar::-webkit-scrollbar-track {
-  background-color: #3c3f41;
-}
-.BSRScrollBar::-webkit-scrollbar-thumb {
-  background-color: #111113;
-  border-radius: 4px;
-}
-
-/* control wrapper */
-#bsr-control-wrapper {
-  display: inline-block;
-  width: 100%;
-}
-#bsr-input-button {
-  float: left;
-}
-#bsr-exit-button {
-  float: right;
-}
-.BSRControlButton {
-  padding: 5px;
-  margin: 3px;
-  display: inline-block;
-}
-
-/* buttons */
-.BSRButton {
-  color: #6b7074;
-  transition: color 500ms ease-out;
-}
-.BSRButton:hover {
-  color: #acaeb1;
-  cursor: pointer;
-}
-
-/* form wrapper, inputs are children of this wrapper */
-#bsr-form-wrapper {
-  padding: 5%;
-}
-
-/* top half of an input */
-.BSRInputTopHalf {
-  display: flex;
-}
-.BSRInputTopHalf > :first-child {
-  border-top-left-radius: 3px;
-  border-bottom-left-radius: 3px;
-}
-.BSRInputTopHalf > :nth-child(2) {
-  border-top-right-radius: 3px;
-  border-bottom-right-radius: 3px;
-}
-
-/* main input, this is where a search term is entered */
-input.BSRMainInputField {
-  background-color: #111113;
-  color: #e8eaed;
-  border: 1px solid #6b7074;
-  height: 25px;
-  width: 200px;
-  margin: 0;
-  padding-left: 5px;
-  box-sizing: border-box;
-}
-input.BSRMainInputField::placeholder {
-  color: #55595d;
-}
-
-/* search modifiers */
-.BSRModifierWrapper {
-  border-top: 1px solid #6b7074;
-  border-right: 1px solid #6b7074;
-  border-bottom: 1px solid #6b7074;
-  height: 25px;
-  min-width: 25px;
-  box-sizing: border-box;
-  position: relative;
-}
-
-/* the element that when hovered shows the modifier dropdown */
-.BSRModifierCoverButton {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-width: 25px;
-  height: 25px;
-}
-.BSRModifierWrapper:hover .BSRModifierDropdown {
-  visibility: visible;
-  max-height: 200px;
-}
-
-/* modifier dropdown */
-.BSRModifierDropdown {
-  visibility: hidden;
-  position: absolute;
-  max-height: 0px;
-  width: 160px;
-  overflow: hidden;
-  left: -45px;
-  top: 23px;
-  border: 1px solid #6b7074;
-  border-radius: 10px;
-  overflow-x: hidden;
-  z-index: 1;
-  transition: max-height 1s, visibility 1s;
-}
-
-/* modifier buttons and inputs */
-.BSRModifierButton {
-  background-color: #202124;
-  padding: 5px;
-  position: relative;
-}
-.BSRModifierButton label:before {
-  position: absolute;
-  content: '';
-  top: 0;
-  bottom: 0;
-  right: 0;
-  left: 0;
-}
-.BSRModifierButton label:hover {
-  cursor: pointer;
-}
-.BSRModifierInput {
-  float: right;
-}
-.BSRModifierButton:hover {
-  background-color: #111113;
-  padding: 5px;
-}
-.BSRMaxMatchLimitWrapper:hover {
-  cursor: default;
-}
-.BSRMaxMatchLimit {
-  width: 50px;
-  height: 25px;
-  background-color: #202124;
-  color: #acaeb1;
-  border: 1px solid #6b7074;
-  border-radius: 3px;
-  box-sizing: border-box;
-}
-.BSRColorPicker::-webkit-color-swatch-wrapper {
-  padding: 0;
-}
-.BSRColorPicker::-webkit-color-swatch {
-  border: none;
-}
-.BSRColorPicker {
-  -webkit-appearance: none;
-  border: none;
-}
-.BSRColorPickerWrapper:hover {
-  cursor: default;
-}
-.BSRColorPicker:hover {
-  cursor: pointer;
-}
-
-/* bottom half of an input */
-.BSRInputBottomHalf {
-  display: flex;
-  padding-top: 5px;
-}
-.BSRActionButton {
-  width: 25px;
-  height: 25px;
-  display: flex;
-  border-radius: 7px;
-  justify-content: center;
-  align-items: center;
-}
-
-.BSRToolTip {
-  background-color: #acaeb1;
-  color: #202124 !important;
-  position: absolute;
-  top: -25px;
-  left: -20px;
-  padding: 2px;
-  border-radius: 3px;
-  opacity: 0;
-  transition: opacity .2s;
-}
-.BSRCopyButton {
-  position: relative;
-}
-`;
