+++ browser-search-regex/chrome/custom_lib/DOM_manipulator/components/Components.ts
@@ -1,0 +1,32 @@
+interface highlightMeOptions {
+  color: string;
+  mods: string;
+  limit: number;
+}
+interface nextMatchOptions {
+  direction: number;
+  newStyles: any;
+  oldStyles: any;
+  scrollBehavior: 'smooth' | 'auto';
+  scrollable: boolean;
+}
+namespace Components {
+  export const queryShadowSelector = (elem: HTMLElement, selector: string) => {
+    const shadow = elem.shadowRoot;
+    if (shadow != null) {
+      const childNodes = Array.from(shadow.childNodes);
+      let parent = childNodes.find((node) => {
+        if (
+          node instanceof HTMLElement &&
+          node.className.includes('shadowWrapper')
+        ) {
+          return node;
+        }
+      });
+      if (parent instanceof HTMLElement && parent != null) {
+        return parent.querySelector(selector);
+      }
+    }
+    return null;
+  };
+}
